requirements.txt
youtube-dl==2021.6.6
tensorflow>=2.6.0
jax>=0.2.21
clu==0.0.6
fabric==2.6.0
tqdm
optax==0.0.9
aioredis==1.3.1
pandas==1.3.0
google-cloud-storage==1.40.0
regex==2021.7.6
scipy==1.7.0
librosa==0.8.1
glob2==0.6
beautifulsoup4==4.10.0
ftfy==6.0.3
scikit-video==1.1.11
numba==0.53.1
numpy==1.19.5
tokenizers==0.10.3
Flask==2.0.2
Flask-Cors==3.0.10
gevent==21.8.0
flax==0.3.4
lxml
tensorflow-datasets==4.4.0
transformers==4.8.2